requirements.txt
gradio
transformers
torch
Pillow
wikipedia
requests
beautifulsoup4
lxml
